refactor: Make MainPlayerManager to execute the functions related to PLAYER
