+++ Assets/2.Scripts/System/DashGhostFx.cs
@@ -22,7 +22,7 @@ public class DashGhostFx : MonoBehaviour
     private void OnEnable()
     {
         _sr = GetComponent<SpriteRenderer>();
-        _player = MainGameManager.Instance.Player.transform;
+        _player = MainPlayerManager.Instance.Player.transform;
 
         _alpha = _alphaSet;
         transform.position = _player.position;
