+++ Assets/2.Scripts/System/main/MainGameManager.cs
@@ -4,44 +4,14 @@ using UnityEngine;
 
 public class MainGameManager : MonoBehaviour
 {
-    public static MainGameManager Instance { get; private set; }
-
-    private void Awake()
-    {
-        Instance = this;
-    }
-
-    //
-    // SINGLETON
-
     [SerializeField]
     private ElevatorMovement _elevatorMovement;
     [SerializeField]
     private CameraFx _cameraFx;
 
-    [SerializeField]
-    private Transform _spawnPos;
-
-    [SerializeField]
-    private GameObject _playerPrefab;
-    private GameObject _player;
-    public GameObject Player
-    {
-        get
-        {
-            return _player;
-        }
-        private set
-        {
-            _player = value;
-        }
-    }
 
     private void Start()
     {
-        // Init player property for develop
-        Player = GameObject.FindGameObjectWithTag("Player");
-
         // Pause all game flow before pressing start button
         MainEventManager.Instance.PauseGamePlayEvent?.Invoke();
     }
@@ -64,8 +34,9 @@ public class MainGameManager : MonoBehaviour
         yield return new WaitForSeconds(2.8f);
 
         // Spawn player
-        Player = Instantiate(_playerPrefab, _spawnPos.position, _spawnPos.rotation);
-        Camera.main.transform.parent.GetComponent<SmoothTargetFollowing>().SetTarget(Player);
+        MainPlayerManager.Instance.SpawnPlayerfromElevator();
+
+        MainPlayerManager.Instance.SetPlayerAsCameraFocus();
 
         // Open elevator door
         yield return StartCoroutine(_elevatorMovement.MoveDoor(true));
