+++ Assets/Resources/Map/B1_mini.psb.meta
@@ -315,7 +315,7 @@ ScriptedImporter:
     spriteName: 
     isGroup: 1
     parentIndex: -1
-    spriteID: 
+    spriteID: 00000000000000000000000000000000
     layerID: -803473935
     mosaicPosition: {x: 0, y: 0}
   - name: "\uC9C0\uC0C19 \uBCF5\uC0AC 2"
