+++ Assets/Resources/Map/sp_mini.psb.meta
@@ -315,7 +315,7 @@ ScriptedImporter:
     spriteName: 
     isGroup: 1
     parentIndex: -1
-    spriteID: 
+    spriteID: 00000000000000000000000000000000
     layerID: -803473935
     mosaicPosition: {x: 0, y: 0}
   - name: "\uB808\uC774\uC5B4 7"
